chore(s2): Cleanup S2 module
