+++ docs/modules/quadkey/api-reference/quadkey.md
@@ -1,3 +1,19 @@
 # quadkey
 
-> TBA
+The quadkey library is currently focused on decoding quadkeys.
+
+> The quadkey functions in math.gl are currently focused on **decoding** S2 encoded data, not encoding it.
+
+## Functions
+
+#### `getQuadkeyLngLat(quadkey: string): number[]`
+
+Returns the center lng, lat of a quadkey cell.
+
+#### `getQuadkeyBoundary(quadkey: string): number[][]`
+
+Returns the center lng, lat of a quadkey cell.
+
+#### `getQuadkeyBoundaryFlat(quadkey: string): number[]`
+
+Returns the center lng, lat of a quadkey cell.
