+++ lerna.json
@@ -1,5 +1,5 @@
 {
-  "version": "3.6.3",
+  "version": "3.7.0-alpha.0",
   "npmClient": "yarn",
   "useWorkspaces": true,
   "packages": [
