+++ modules/quadkey/test/quadkey.spec.ts
@@ -1,5 +1,5 @@
 import test from 'tape-promise/tape';
-import {quadkeyToWorldBounds, getQuadkeyPolygon} from '@math.gl/quadkey';
+import {quadkeyToWorldBounds, getQuadkeyBoundary} from '@math.gl/quadkey';
 
 const TEST_DATA = [
   {
@@ -34,9 +34,9 @@ test('quadkey#quadkeyToWorldBounds', (t) => {
   t.end();
 });
 
-test('quadkey#getQuadkeyPolygon', (t) => {
+test('quadkey#getQuadkeyBoundary', (t) => {
   for (const {quadkey} of TEST_DATA) {
-    const polygon = getQuadkeyPolygon(quadkey);
+    const polygon = getQuadkeyBoundary(quadkey);
     t.ok(polygon instanceof Array, 'polygon is flat array');
     t.is(polygon.length / 2 - 1, 4, 'polygon has 4 sides');
     t.deepEqual(polygon.slice(0, 2), polygon.slice(-2), 'polygon is closed');
