+++ modules/s2/src/index.ts
@@ -1,13 +1,12 @@
 // math.gl MIT license
 
-export {getIdFromToken, getGeoBounds, getS2QuadKey, getS2Polygon} from './s2-utils';
+export {getS2IndexFromToken, getS2TokenFromIndex, getS2ChildIndex} from './s2-token-functions';
+export {getS2BoundaryFlat} from './s2-geometry-functions';
 
-export {
-  getTokenFromId,
-  getS2ChildCellId,
-  getS2CellFromToken,
-  get2dRegionFromS2Cell
-} from './s2-utils-ext';
+export {getS2Cell as _getS2Cell} from './s2geometry/s2-cell-utils';
 
-export type {S2HeightInfo} from './s2-to-obb-points';
-export {getOrientedBoundingBoxCornerPoints} from './s2-to-obb-points';
+export {getS2GeoBounds as _getS2GeoBounds} from './converters/s2-to-boundary';
+export {getS2Region as _getS2Region} from './converters/s2-to-region';
+
+export type {S2HeightInfo} from './converters/s2-to-obb-points';
+export {getS2OrientedBoundingBoxCornerPoints as _getS2OrientedBoundingBoxCornerPoints} from './converters/s2-to-obb-points';
