+++ modules/s2/test/index.ts
@@ -1,1 +1,2 @@
+import './s2-geometry.spec';
 import './s2-utils.spec';
