+++ test/modules.spec.ts
@@ -1,13 +1,13 @@
-import '../modules/types/test';
-import '../modules/core/test';
-import '../modules/culling/test';
-import '../modules/geoid/test';
-import '../modules/geospatial/test';
-import '../modules/polygon/test';
-import '../modules/proj4/test';
-import '../modules/sun/test';
-import '../modules/web-mercator/test';
+// import '../modules/types/test';
+// import '../modules/core/test';
+// import '../modules/culling/test';
+// import '../modules/geoid/test';
+// import '../modules/geospatial/test';
+// import '../modules/polygon/test';
+// import '../modules/proj4/test';
+// import '../modules/sun/test';
+// import '../modules/web-mercator/test';
 
-import '../modules/geohash/test';
-// import '../modules/s2/test';
-import '../modules/quadkey/test';
+// import '../modules/geohash/test';
+// import '../modules/quadkey/test';
+import '../modules/s2/test';
